+++ salt/tests/unit/formulas/fixtures/salt.py
@@ -424,6 +424,9 @@ register_basic("metalk8s_network.get_control_plane_ingress_ip")(
 register_basic("metalk8s_network.get_control_plane_ingress_endpoint")(
     MagicMock(return_value="https://192.168.1.240:8443")
 )
+register_basic("metalk8s_network.get_portmap_ips")(
+    MagicMock(return_value=["192.168.1.100", "127.0.0.1"])
+)
 
 
 @register_basic("metalk8s_network.get_ip_from_cidrs")